Move (CsvFileExportPrinter), refs 2740
